re added styles
